Lab7 work

+++ 421/app_api/controllers/ctrlBlog.js
@@ -1,4 +1,5 @@
-const Blog = require('../../models/blogs');
+const Blog = require('../../../../WebDev/421/models/blogs');
+const jwt = require('jsonwebtoken');
 
 exports.home = function(req,res) {
     res.render('home', { title: 'Patrick McCullough Blog'});
@@ -6,7 +7,7 @@ exports.home = function(req,res) {
 // Get all blogs
 module.exports.getAllBlogs = async (req, res) => {
     try {
-        const blogs = await Blog.find();
+        const blogs = await Blog.find().populate('blogAuthor', 'name');
         res.json(blogs); // Return blogs as JSON
     } catch (error) {
         res.status(500).json({ error: 'Internal Server Error' });
@@ -16,7 +17,7 @@ module.exports.getAllBlogs = async (req, res) => {
 // Get a single blog by ID
 module.exports.getBlogById = async (req, res) => {
     try {
-        const blog = await Blog.findById(req.params.id);
+        const blog = await Blog.findById(req.params.id).populate('blogAuthor', 'name');
         if (!blog) {
             return res.status(404).json({ error: 'Blog not found' });
         }
@@ -29,35 +30,57 @@ module.exports.getBlogById = async (req, res) => {
 // Add a new blog
 module.exports.createBlog = async (req, res) => {
     try {
-        const newBlog = new Blog(req.body);
+        const newBlog = new Blog({
+            blogTitle: req.body.blogTitle,
+            blogText: req.body.blogText,
+            blogAuthor: req.userId // Ensure this matches the decoded token
+        });
+    
         const savedBlog = await newBlog.save();
-        res.status(201).json(savedBlog); // Return the saved blog
+        res.status(201).json(savedBlog);
     } catch (error) {
+        console.log("Error in creating blog:", error);
         res.status(400).json({ error: 'Failed to add blog' });
     }
 };
 
 
 module.exports.updateBlog = async (req, res) => {
+    const blogId = req.params.id;
     try {
-        const blogId = req.params.id;
-        const updatedBlog = await Blog.findByIdAndUpdate(blogId, req.body, { new: true });
-        if (!updatedBlog) {
+        const blog = await Blog.findById(blogId);
+        if (!blog) {
             return res.status(404).json({ error: 'Blog not found' });
         }
-        res.json(updatedBlog); // Return the updated blog
+
+        if (blog.blogAuthor.toString() !== req.userId) {
+            return res.status(403).json({ error: 'Unauthorized to update this blog' });
+        }
+        const updatedBlog = await Blog.findByIdAndUpdate(blogId, req.body, { new: true });
+        res.json(updatedBlog);
     } catch (error) {
-        res.status(500).json({ error: 'Error updating the blog' });
+        console.error("Error updating blog:", error);
+        res.status(500).json({ error: 'Error updating the blog: ' + error.message });
     }
 };
 
 
-    module.exports.deleteBlog = async (req, res) => {
-        try {
-            const blogId = req.params.id;
-            await Blog.findByIdAndDelete(blogId);
-            res.json({ message: 'Blog successfully deleted' });
-        } catch (error) {
-            res.status(500).json({ error: 'Error deleting the blog' });
+module.exports.deleteBlog = async (req, res) => {
+    const blogId = req.params.id;
+    try {
+        const blog = await Blog.findById(blogId);
+        if (!blog) {
+            return res.status(404).json({ error: 'Blog not found' });
         }
-    };
+        if (blog.blogAuthor.toString() !== req.userId) {
+            return res.status(403).json({ error: 'Unauthorized to delete this blog' });
+        }
+        await Blog.findByIdAndDelete(blogId);
+        console.log("Blog successfully deleted", blogId);  // Confirm deletion in logs
+        res.json({ message: 'Blog successfully deleted' });
+    } catch (error) {
+        console.error("Error deleting blog:", error);
+        res.status(500).json({ error: 'Error deleting the blog: ' + error.message });
+    }
+
+};
